refactor: move methods between packages
