+++ packages/api/package.json
@@ -21,7 +21,6 @@
   },
   "dependencies": {
     "@stacks/common": "^6.8.1",
-    "@stacks/transactions": "^6.8.1",
     "cross-fetch": "^3.1.5"
   },
   "devDependencies": {
