+++ packages/api/src/api.ts
@@ -6,9 +6,17 @@ import {
   bytesToHex,
   Hex,
   validateHash256,
+  ClarityAbi,
 } from '@stacks/common';
 import { createFetchFn, FetchFn } from './fetch';
-import { TxBroadcastResult, TxBroadcastResultOk, TxBroadcastResultRejected } from './types';
+import {
+  FeeEstimateResponse,
+  FeeEstimation,
+  NoEstimateAvailableError,
+  TxBroadcastResult,
+  TxBroadcastResultOk,
+  TxBroadcastResultRejected,
+} from './types';
 
 export const HIRO_MAINNET_URL = 'https://stacks-node-api.mainnet.stacks.co';
 export const HIRO_TESTNET_URL = 'https://stacks-node-api.testnet.stacks.co';
@@ -20,11 +28,6 @@ export const TRANSACTION_FEE_ESTIMATE_PATH = '/v2/fees/transaction';
 export const ACCOUNT_PATH = '/v2/accounts';
 export const CONTRACT_ABI_PATH = '/v2/contracts/interface';
 export const READONLY_FUNCTION_CALL_PATH = '/v2/contracts/call-read';
-
-/** @ignore internal */
-export const StacksNetworks = ['mainnet', 'testnet', 'devnet', 'mocknet'] as const;
-/** The enum-style names of different common Stacks networks */
-export type StacksNetworkName = (typeof StacksNetworks)[number];
 
 /**
  * todo
@@ -81,47 +84,41 @@ export class StacksApi {
     return broadcastTransaction({ tx, attachment, url: this.url });
   };
 
-  getTransferFeeEstimateApiUrl = () => `${this.url}${this.transferFeeEstimateEndpoint}`;
-  getTransactionFeeEstimateApiUrl = () => `${this.url}${this.transactionFeeEstimateEndpoint}`;
-  getAccountApiUrl = (address: string) => `${this.url}${this.accountEndpoint}/${address}?proof=0`;
-  getAccountExtendedBalancesApiUrl = (address: string) =>
-    `${this.url}/extended/v1/address/${address}/balances`;
-  getAbiApiUrl = (address: string, contract: string) =>
-    `${this.url}${this.contractAbiEndpoint}/${address}/${contract}`;
-  getReadOnlyFunctionCallApiUrl = (
-    contractAddress: string,
-    contractName: string,
-    functionName: string
-  ) =>
-    `${this.url}${
-      this.readOnlyFunctionCallEndpoint
-    }/${contractAddress}/${contractName}/${encodeURIComponent(functionName)}`;
-  getInfoUrl = () => `${this.url}/v2/info`;
-  getBlockTimeInfoUrl = () => `${this.url}/extended/v1/info/network_block_times`;
-  getPoxInfoUrl = () => `${this.url}/v2/pox`;
-  getRewardsUrl = (address: string, options?: any) => {
-    let url = `${this.url}/extended/v1/burnchain/rewards/${address}`;
-    if (options) {
-      url = `${url}?limit=${options.limit}&offset=${options.offset}`;
-    }
-    return url;
+  /**
+   * Lookup the nonce for an address from a core node
+   * @param address - The Stacks address to look up the next nonce for
+   * @return A promise that resolves to a bigint of the next nonce
+   */
+  getNonce = async (address: string): Promise<bigint> => {
+    return getNonce({ address, url: this.url });
   };
-  getRewardsTotalUrl = (address: string) =>
-    `${this.url}/extended/v1/burnchain/rewards/${address}/total`;
-  getRewardHoldersUrl = (address: string, options?: any) => {
-    let url = `${this.url}/extended/v1/burnchain/reward_slot_holders/${address}`;
-    if (options) {
-      url = `${url}?limit=${options.limit}&offset=${options.offset}`;
-    }
-    return url;
+
+  /**
+   * Estimate the total transaction fee in microstacks for a Stacks transaction
+   * @param payload - The transaction to estimate fees for
+   * @param estimatedLength - Optional argument that provides the endpoint with an
+   * estimation of the final length (in bytes) of the transaction, including any post-conditions
+   * and signatures
+   * @return A promise that resolves to an array of {@link FeeEstimate}
+   */
+  estimateTransaction = async (
+    payload: Hex,
+    estimatedLength?: number
+  ): Promise<[FeeEstimation, FeeEstimation, FeeEstimation]> => {
+    return estimateTransaction({ payload, estimatedLength, url: this.url });
   };
-  getStackerInfoUrl = (contractAddress: string, contractName: string) =>
-    `${this.url}${this.readOnlyFunctionCallEndpoint}
-    ${contractAddress}/${contractName}/get-stacker-info`;
-  getDataVarUrl = (contractAddress: string, contractName: string, dataVarName: string) =>
-    `${this.url}/v2/data_var/${contractAddress}/${contractName}/${dataVarName}?proof=0`;
-  getMapEntryUrl = (contractAddress: string, contractName: string, mapName: string) =>
-    `${this.url}/v2/map_entry/${contractAddress}/${contractName}/${mapName}?proof=0`;
+
+  /**
+   * Fetch a contract's ABI
+   * @param address - The contracts address
+   * @param contractName - The contracts name
+   * @returns A promise that resolves to a ClarityAbi if the operation succeeds
+   */
+  getAbi = async (address: string, contractName: string): Promise<ClarityAbi> => {
+    return getAbi({ address, contractName, url: this.url });
+  };
+
+  // todo: migrate to new api pattern
   getNameInfo(fullyQualifiedName: string) {
     /*
       TODO: Update to v2 API URL for name lookups
@@ -150,9 +147,18 @@ export class StacksApi {
   }
 }
 
+export type ApiBaseParams = {
+  /** Optional API/Node base URL, defaults to {@link HIRO_MAINNET_URL} */
+  url?: string;
+  /** Optional fetch function, defaults to new instance of {@link createFetchFn} */
+  fetchFn?: FetchFn;
+};
+
 /**
  * Broadcast a serialized transaction to a Stacks node (which will validate and forward to the network).
- * @returns a Promise that resolves to a {@link TxBroadcastResult} object
+ * @param opts.tx - A serialized transaction as a hex string
+ * @param opts.attachment - Optional attachment encoded as a hex string
+ * @returns A Promise that resolves to a {@link TxBroadcastResult} object
  */
 async function broadcastTransaction({
   tx,
@@ -164,11 +170,7 @@ async function broadcastTransaction({
   tx: Hex;
   /** Optional attachment encoded as a hex string */
   attachment?: Hex;
-  /** Optional API/Node base URL, defaults to {@link HIRO_MAINNET_URL} */
-  url?: string;
-  /** Optional fetch function, defaults to new instance of {@link createFetchFn} */
-  fetchFn?: FetchFn;
-}): Promise<TxBroadcastResult> {
+} & ApiBaseParams): Promise<TxBroadcastResult> {
   const json = attachment ? { tx, attachment } : { tx };
   const options = {
     method: 'POST',
@@ -190,4 +192,134 @@ async function broadcastTransaction({
   if (!validateHash256(txid)) throw new Error(text);
 
   return { txid } as TxBroadcastResultOk;
+}
+
+/**
+ * Lookup the nonce for an address from a core node
+ * @param opts.address - The Stacks address to look up the next nonce for
+ * @return A promise that resolves to an integer
+ */
+export async function getNonce({
+  address,
+  url = HIRO_MAINNET_URL,
+  fetchFn = createFetchFn(),
+}: {
+  /** The Stacks address to look up the next nonce for */
+  address: string;
+} & ApiBaseParams): Promise<bigint> {
+  const response = await fetchFn(`${url}${ACCOUNT_PATH}/${address}?proof=0`);
+  if (!response.ok) {
+    const msg = await response.text().catch(() => '');
+    throw new Error(
+      `Error fetching nonce. Response ${response.status}: ${response.statusText}. Attempted to fetch ${url} and failed with the message: "${msg}"`
+    );
+  }
+  const responseText = await response.text();
+  const result = JSON.parse(responseText) as { nonce: string };
+  return BigInt(result.nonce);
+}
+
+/**
+ * @deprecated Use the new {@link estimateTransaction} function instead.
+ *
+ * Estimate the total transaction fee in microstacks for a token transfer
+ *
+ * ⚠ Only sensible for token transfer transactions!
+ * @param opts.tx - The token transfer transaction to estimate fees for
+ * @return A promise that resolves to number of microstacks per byte
+ */
+export async function estimateTransfer({
+  tx,
+  url = HIRO_MAINNET_URL,
+  fetchFn = createFetchFn(),
+}: {
+  /** The token transfer transaction to estimate fees for */
+  tx: Hex;
+} & ApiBaseParams): Promise<bigint> {
+  const response = await fetchFn(`${url}${TRANSFER_FEE_ESTIMATE_PATH}`, {
+    headers: {
+      Accept: 'application/text',
+    },
+  });
+
+  if (!response.ok) {
+    const msg = await response.text().catch(() => '');
+    throw new Error(
+      `Error estimating transfer fee. Response ${response.status}: ${response.statusText}. Attempted to fetch ${url} and failed with the message: "${msg}"`
+    );
+  }
+  const feeRateResult = await response.text();
+  const txBytes = BigInt(Math.ceil(tx.length / 2));
+  const feeRate = BigInt(feeRateResult);
+  return feeRate * txBytes;
+}
+
+/**
+ * Estimate the total transaction fee in microstacks for a Stacks transaction
+ * @param opts.payload - The transaction to estimate fees for
+ * @param opts.estimatedLength - Optional argument that provides the endpoint with an
+ * estimation of the final length (in bytes) of the transaction, including any post-conditions
+ * and signatures
+ * @return A promise that resolves to FeeEstimate
+ */
+export async function estimateTransaction({
+  payload,
+  estimatedLength,
+  url = HIRO_MAINNET_URL,
+  fetchFn = createFetchFn(),
+}: {
+  payload: Hex;
+  estimatedLength?: number;
+} & ApiBaseParams): Promise<[FeeEstimation, FeeEstimation, FeeEstimation]> {
+  const json = {
+    transaction_payload: payload,
+    estimated_len: estimatedLength,
+  };
+  const options = {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify(json),
+  };
+
+  const response = await fetchFn(`${url}${TRANSACTION_FEE_ESTIMATE_PATH}`, options);
+  if (!response.ok) {
+    const body = await response.json().catch(() => ({}));
+
+    if (body?.reason === 'NoEstimateAvailable') {
+      throw new NoEstimateAvailableError(body?.reason_data?.message ?? '');
+    }
+
+    throw new Error(
+      `Error estimating transaction fee. Response ${response.status}: ${response.statusText}. Attempted to fetch ${url} and failed with the message: "${body}"`
+    );
+  }
+
+  const data: FeeEstimateResponse = await response.json();
+  return data.estimations;
+}
+
+/**
+ * Fetch a contract's ABI
+ * @param opts.address - The contracts address
+ * @param opts.contractName - The contracts name
+ * @returns A promise that resolves to a ClarityAbi if the operation succeeds
+ */
+export async function getAbi({
+  address,
+  contractName,
+  url = HIRO_MAINNET_URL,
+  fetchFn = createFetchFn(),
+}: {
+  address: string;
+  contractName: string;
+} & ApiBaseParams): Promise<ClarityAbi> {
+  const response = await fetchFn(`${url}${CONTRACT_ABI_PATH}/${address}/${contractName}`);
+  if (!response.ok) {
+    const msg = await response.text().catch(() => '');
+    throw new Error(
+      `Error fetching contract ABI for contract "${contractName}" at address ${address}. Response ${response.status}: ${response.statusText}. Attempted to fetch ${url} and failed with the message: "${msg}"`
+    );
+  }
+
+  return JSON.parse(await response.text()) as ClarityAbi;
 }
