+++ packages/transactions/src/builders.ts
@@ -70,201 +70,6 @@ import { StacksTransaction } from './transaction';
 import { createLPList } from './types';
 import { cvToHex, omit, parseReadOnlyResponse, validateTxId } from './utils';
 
-/**
- * Lookup the nonce for an address from a core node
- *
- * @param {string} address - the c32check address to look up
- * @param {StacksNetworkName | StacksNetwork} network - the Stacks network to look up address on
- *
- * @return a promise that resolves to an integer
- */
-export async function getNonce(
-  address: string,
-  network?: StacksNetworkName | StacksNetwork
-): Promise<bigint> {
-  const derivedNetwork = StacksNetwork.fromNameOrNetwork(network ?? new StacksMainnet());
-  const url = derivedNetwork.getAccountApiUrl(address);
-
-  const response = await derivedNetwork.fetchFn(url);
-  if (!response.ok) {
-    let msg = '';
-    try {
-      msg = await response.text();
-    } catch (error) {}
-    throw new Error(
-      `Error fetching nonce. Response ${response.status}: ${response.statusText}. Attempted to fetch ${url} and failed with the message: "${msg}"`
-    );
-  }
-  const responseText = await response.text();
-  const result = JSON.parse(responseText) as { nonce: string };
-  return BigInt(result.nonce);
-}
-
-/**
- * @deprecated Use the new {@link estimateTransaction} function instead.
- *
- * Estimate the total transaction fee in microstacks for a token transfer
- *
- * @param {StacksTransaction} transaction - the token transfer transaction to estimate fees for
- * @param {StacksNetworkName | StacksNetwork} network - the Stacks network to estimate transaction for
- *
- * @return a promise that resolves to number of microstacks per byte
- */
-export async function estimateTransfer(
-  transaction: StacksTransaction,
-  network?: StacksNetworkName | StacksNetwork
-): Promise<bigint> {
-  if (transaction.payload.payloadType !== PayloadType.TokenTransfer) {
-    throw new Error(
-      `Transaction fee estimation only possible with ${
-        PayloadType[PayloadType.TokenTransfer]
-      } transactions. Invoked with: ${PayloadType[transaction.payload.payloadType]}`
-    );
-  }
-
-  return estimateTransferUnsafe(transaction, network);
-}
-
-/**
- * @deprecated Use the new {@link estimateTransaction} function instead.
- * @internal
- */
-export async function estimateTransferUnsafe(
-  transaction: StacksTransaction,
-  network?: StacksNetworkName | StacksNetwork
-): Promise<bigint> {
-  const requestHeaders = {
-    Accept: 'application/text',
-  };
-
-  const fetchOptions = {
-    method: 'GET',
-    headers: requestHeaders,
-  };
-
-  const derivedNetwork = StacksNetwork.fromNameOrNetwork(network ?? deriveNetwork(transaction));
-  const url = derivedNetwork.getTransferFeeEstimateApiUrl();
-
-  const response = await derivedNetwork.fetchFn(url, fetchOptions);
-  if (!response.ok) {
-    let msg = '';
-    try {
-      msg = await response.text();
-    } catch (error) {}
-    throw new Error(
-      `Error estimating transaction fee. Response ${response.status}: ${response.statusText}. Attempted to fetch ${url} and failed with the message: "${msg}"`
-    );
-  }
-  const feeRateResult = await response.text();
-  const txBytes = BigInt(transaction.serialize().byteLength);
-  const feeRate = BigInt(feeRateResult);
-  return feeRate * txBytes;
-}
-
-interface FeeEstimation {
-  fee: number;
-  fee_rate: number;
-}
-interface FeeEstimateResponse {
-  cost_scalar_change_by_byte: bigint;
-  estimated_cost: {
-    read_count: bigint;
-    read_length: bigint;
-    runtime: bigint;
-    write_count: bigint;
-    write_length: bigint;
-  };
-  estimated_cost_scalar: bigint;
-  estimations: [FeeEstimation, FeeEstimation, FeeEstimation];
-}
-
-/**
- * Estimate the total transaction fee in microstacks for a Stacks transaction
- *
- * @param {StacksTransaction} transaction - the transaction to estimate fees for
- * @param {number} estimatedLen - is an optional argument that provides the endpoint with an
- * estimation of the final length (in bytes) of the transaction, including any post-conditions
- * and signatures
- * @param {StacksNetworkName | StacksNetwork} network - the Stacks network to estimate transaction fees for
- *
- * @return a promise that resolves to FeeEstimate
- */
-export async function estimateTransaction(
-  transactionPayload: Payload,
-  estimatedLen?: number,
-  network?: StacksNetworkName | StacksNetwork
-): Promise<[FeeEstimation, FeeEstimation, FeeEstimation]> {
-  const options = {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({
-      transaction_payload: bytesToHex(serializePayload(transactionPayload)),
-      ...(estimatedLen ? { estimated_len: estimatedLen } : {}),
-    }),
-  };
-
-  const derivedNetwork = StacksNetwork.fromNameOrNetwork(network ?? new StacksMainnet());
-  const url = derivedNetwork.getTransactionFeeEstimateApiUrl();
-
-  const response = await derivedNetwork.fetchFn(url, options);
-
-  if (!response.ok) {
-    const body = await response.json().catch(() => ({}));
-
-    if (body?.reason === 'NoEstimateAvailable') {
-      throw new NoEstimateAvailableError(body?.reason_data?.message ?? '');
-    }
-
-    throw new Error(
-      `Error estimating transaction fee. Response ${response.status}: ${response.statusText}. Attempted to fetch ${url} and failed with the message: "${body}"`
-    );
-  }
-
-  const data: FeeEstimateResponse = await response.json();
-  return data.estimations;
-}
-
-/**
- * Fetch a contract's ABI
- *
- * @param {string} address - the contracts address
- * @param {string} contractName - the contracts name
- * @param {StacksNetworkName | StacksNetwork} network - the Stacks network to broadcast transaction to
- *
- * @returns {Promise} that resolves to a ClarityAbi if the operation succeeds
- */
-export async function getAbi(
-  address: string,
-  contractName: string,
-  network: StacksNetworkName | StacksNetwork
-): Promise<ClarityAbi> {
-  const options = {
-    method: 'GET',
-  };
-
-  const derivedNetwork = StacksNetwork.fromNameOrNetwork(network);
-  const url = derivedNetwork.getAbiApiUrl(address, contractName);
-
-  const response = await derivedNetwork.fetchFn(url, options);
-  if (!response.ok) {
-    const msg = await response.text().catch(() => '');
-    throw new Error(
-      `Error fetching contract ABI for contract "${contractName}" at address ${address}. Response ${response.status}: ${response.statusText}. Attempted to fetch ${url} and failed with the message: "${msg}"`
-    );
-  }
-
-  return JSON.parse(await response.text()) as ClarityAbi;
-}
-
-function deriveNetwork(transaction: StacksTransaction) {
-  switch (transaction.version) {
-    case TransactionVersion.Mainnet:
-      return new StacksMainnet();
-    case TransactionVersion.Testnet:
-      return new StacksTestnet();
-  }
-}
-
 export interface MultiSigOptions {
   numSignatures: number;
   publicKeys: string[];
